SCC/final/requirements.txt
appnope==0.1.0
astroid==2.3.2
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
bleach==3.1.5
cffi==1.14.0
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dnspython==1.16.0
entrypoints==0.3
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
ipykernel==5.2.1
ipynb-py-convert==0.4.5
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
meteocalc==1.1.0
mistune==0.8.4
mlxtend==0.17.2
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.1
p2j==1.3.2
packaging==20.3
paho-mqtt==1.5.0
pandas==1.0.1
pandocfilters==1.4.2
parsel==1.5.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
Protego==0.1.16
psycopg2==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.1
pylint==2.4.3
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySide2==5.14.2.1
pyt==1.0.5
python-dateutil==2.8.1
python-etcd==0.4.5
pytz==2019.3
pyzmq==19.0.1
queuelib==1.5.0
scikit-learn==0.23.0
scipy==1.4.1
Scrapy==1.8.0
scrapy-rotating-proxies==0.6.2
Scrapy-UserAgents==0.0.1
Send2Trash==1.5.0
service-identity==18.1.0
shiboken2==5.14.2.1
six==1.14.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.0.0
tornado==6.0.4
traitlets==4.3.3
Twisted==19.10.0
typed-ast==1.4.0
typing==3.7.4.1
urllib3==1.25.9
w3lib==1.21.0
wcwidth==0.1.9
webencodings==0.5.1
wrapt==1.11.2
zipp==3.1.0
zope.interface==4.7.1
